requirements_v4.txt
git+https://github.com/botswana-harvard/django-crypto-fields.git@0.1.14#egg=django-crypto-fields
git+https://github.com/botswana-harvard/django-revision.git@0.1.14#egg=django-revision
git+https://github.com/botswana-harvard/edc-action-item.git@develop#egg=edc_action_item
git+https://github.com/botswana-harvard/edc-appointment.git@develop#egg=edc_appointment 
git+https://github.com/botswana-harvard/edc-base.git@develop#egg=edc_base
git+https://github.com/botswana-harvard/edc-call-manager.git@develop#egg=edc-call-manager
git+https://github.com/botswana-harvard/edc-calendar.git@develop#egg=edc-calendar
git+https://github.com/botswana-harvard/edc-constants.git@develop#egg=edc_constants
git+https://github.com/botswana-harvard/edc-consent.git@develop#egg=edc_consent
git+https://github.com/botswana-harvard/edc-dashboard.git@develop#egg=edc_dashboard
git+https://github.com/botswana-harvard/edc-data-manager.git@0.1.5#egg=edc-data-manager
git+https://github.com/botswana-harvard/edc-device.git@develop#egg=edc_device
git+https://github.com/botswana-harvard/edc-document-archieve-api.git@develop
git+https://github.com/botswana-harvard/edc-facility.git@develop#egg=edc_facility
git+https://github.com/botswana-harvard/edc-fieldsets.git@develop#egg=edc_fieldsets
git+https://github.com/botswana-harvard/edc-form-validators.git@develop#egg=edc_form_validators
git+https://github.com/botswana-harvard/edc-identifier.git@develop#egg=edc_identifier
git+https://github.com/botswana-harvard/edc-lab.git@develop#egg=edc_lab
git+https://github.com/botswana-harvard/edc-label.git@develop#egg=edc_label
git+https://github.com/botswana-harvard/edc-locator.git@develop#egg=edc_locator
git+https://github.com/botswana-harvard/edc-list-data.git@develop#egg=edc_list_data
git+https://github.com/botswana-harvard/edc-metadata.git@develop#egg=edc_metadata
git+https://github.com/botswana-harvard/edc-metadata-rules.git@develop#egg=edc_metadata_rules
git+https://github.com/botswana-harvard/edc-model-admin.git@develop#egg=edc_model_admin
git+https://github.com/botswana-harvard/edc-model-wrapper.git@develop#egg=edc_model_wrapper
git+https://github.com/botswana-harvard/edc-navbar.git@develop#egg=edc_navbar
git+https://github.com/botswana-harvard/edc-odk.git@develop#egg=edc-odk
git+https://github.com/botswana-harvard/edc-prn.git@develop#egg=edc_prn
git+https://github.com/botswana-harvard/edc-protocol.git@develop#egg=edc_protocol
git+https://github.com/botswana-harvard/edc-reference.git@develop#egg=edc_reference
git+https://github.com/botswana-harvard/edc-registration.git@develop#egg=edc_registration
git+https://github.com/botswana-harvard/edc-reports.git@develop#egg=edc_reports
git+https://github.com/botswana-harvard/edc-search.git@develop#egg=edc_search
git+https://github.com/botswana-harvard/edc-senaite-interface.git@develop#egg=edc-senaite-interface
git+https://github.com/botswana-harvard/edc-sms.git@develop#egg=edc_sms
git+https://github.com/botswana-harvard/edc-subject-dashboard.git@develop#egg=edc_subject_dashboard
git+https://github.com/botswana-harvard/edc-timepoint.git@develop#egg=edc_timepoint
git+https://github.com/botswana-harvard/edc-visit-schedule.git@develop#egg=edc_visit_schedule
git+https://github.com/botswana-harvard/edc-visit-tracking.git@develop#egg=edc_visit_tracking
git+https://github.com/flourishbhp/flourish-child.git@consent-v4#egg=flourish_child
git+https://github.com/flourishbhp/flourish-child-validations.git@consent-v4#egg=flourish_child_validations
git+https://github.com/flourishbhp/flourish-caregiver.git@consent-v4#egg=flourish_caregiver
git+https://github.com/flourishbhp/flourish-dashboard.git@consent-v4#egg=flourish_dashboard
git+https://github.com/flourishbhp/flourish-visit-schedule.git@consent-v4#egg=flourish_visit_schedule
git+https://github.com/flourishbhp/flourish-follow.git@develop#egg=flourish_follow
git+https://github.com/flourishbhp/flourish-reports.git@develop#egg=flourish_reports
git+https://github.com/flourishbhp/flourish-form-validations.git@consent-v4#egg=flourish_form_validations
git+https://github.com/flourishbhp/flourish-prn.git@develop#egg=flourish_prn
git+https://github.com/flourishbhp/flourish-export.git@develop#egg=flourish_export
git+https://github.com/flourishbhp/pre-flourish.git@develop#egg=pre_flourish
git+https://github.com/flourishbhp/flourish-reference@consent-v4#egg=flourish_reference
git+https://github.com/flourishbhp/flourish-metadata-rules@consent-v4#egg=flourish_metadata_rules
git+https://github.com/flourishbhp/flourish-labs@develop#egg=flourish_labs
git+https://github.com/flourishbhp/flourish-calendar@develop#egg=flourish_calendar
git+https://github.com/flourishbhp/pre-flourish-follow.git@develop#egg=pre_flourish_follow
git+https://github.com/flourishbhp/flourish-facet.git@develop#egg=flourish_facet
pypdfium2==3.8.0
django_tables2
django-crispy-forms==1.14.0
